Added a new type of learning rate: LearningErrorRate.
Based off of min-rate newbob, this learning rate will be scaled back when the validation error does not decrease. The difference is that instead of waiting a set number of epochs before the scale factor can be applied, the scale factor can only be applied when the validation error reaches a certain threshold (eg If the validation error does not decrease and it is less than 20%, scale learning rate).

This is useful for running hyperparameter sweeps, where different tested architectures take a wide range of epochs to converge. Defining a static number of epochs to scale back learning rate can cut short training on models that take longer to train, or overtrain models that trian quickly.